+++ changes.md
@@ -1,7 +1,8 @@
-**This build requires [vilib v1.1.0](https://github.com/Efnilite/vilib/releases/latest) and [IP v4.1.1](https://github.com/Efnilite/Walk-in-the-Park/releases/tag/v4.1.1)✨**
+**This build requires [vilib v1.1.0](https://github.com/Efnilite/vilib/releases/latest) and [IP v5.0.0](https://github.com/Efnilite/Walk-in-the-Park/releases/tag/v4.1.1)✨**
 
 - Added invite feedback
 - Added Hourglass time option to config
+- Added reload command (/ipp reload)
 - Fixed trial times being able to be registered when score was below goal
 - Fixed Duels player cap, you can now have infinite players
 - Fixed Duels not supporting > 4 players